requirements.txt
beautifulsoup4==4.12.2
easygui==0.98.3
image==1.5.33
pandas==2.0.0
Pillow==9.4.0
requests==2.28.2
